Ejercicios

+++ Boletin1/src/Ejercicio7.java
@@ -5,13 +5,13 @@ public class Ejercicio7 {
         System.out.println("Introduzca la cantidad de numeros que va a introducir: ");
         Scanner sc = new Scanner(System.in);
         int cant = sc.nextInt();
-        int suma = 0;
-        for (int i = 1; i <= cant ; i++) {
-            System.out.println("Introduce el "+i+"º numero");
+        double suma = 0;
+        for (int i = 1; i <= cant; i++) {
+            System.out.println("Introduce el " + i + "º numero");
             int num = sc.nextInt();
             suma += num;
         }
-        double media =suma/cant;
-        System.out.println("La media de esos numeros es "+ media);
+        double media = suma / cant;
+        System.out.println("La media de esos numeros es " + media);
     }
 }
